requirements.txt
bleach==2.1.3
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
decorator==4.2.1
Django==2.0.3
entrypoints==0.2.3
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
html5lib==1.0.1
httplib2==0.11.0
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
oauth2client==4.1.2
oauthlib==2.0.7
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.7.1
pytz==2018.3
pywinpty==0.5.1
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
